requirements.txt
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.0
Pillow==11.3.0
django-cors-headers==4.7.0